backend/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-multipart>=0.0.21
google-generativeai>=0.8.0
pydantic>=2.11.0
python-dotenv>=0.19.0
